Refactor AST and interpreter to improve type handling and function evaluation
- Add `statementNode()` and `expressionNode()` methods to AST nodes for type safety
- Enhance function definition and call parsing with more robust type checking
- Update environment to support type-annotated variable declarations
- Improve error handling for type mismatches in function calls and assignments
- Remove deprecated keywords like 'let' and 'var' from lexer
- Simplify function parameter parsing and type annotation requirements

+++ interpreter/interpreter_test.go
@@ -248,7 +248,7 @@ func testIntegerValue(t *testing.T, obj Value, expected int) bool {
 		t.Errorf("object is not IntegerValue. got=%T (%+v)", obj, obj)
 		return false
 	}
-	if result.Value != expected {
+	if result.Value != int64(expected) {
 		t.Errorf("object has wrong value. got=%d, want=%d", result.Value, expected)
 		return false
 	}
